Wait to let nodes exchange gossip
After the recent changes trusted peers are not inserted
in the p2p topology at startup. Instaed the node connect
to them at startup and then they will exchange gossips as usual.
This means that the nodes needs some time (a few secs or less) to
insert each other in their respective topology.